fix(LP64): fix pass size_t/offset_t to printf() format width specifier which accepts int
This causes issues when compiling on platforms with LP64 data model.

* maple/menu.c menu_foot()
* so/chat.c chat_topic()
* maple/more.c outs_footer()

+++ maple/menu.c
@@ -412,7 +412,7 @@ menu_foot(void)
     sprintf(footer, "\x1b[0;34;46m%s%d:%02d] \x1b[30;47m 目前站上有\x1b[31m%4d\x1b[30m 人，我是 \x1b[31m%-*s\t\x1b[30m [呼叫/訊息]\x1b[31m%s",
         datemsg, ufo / 60, ufo % 60,
         /*ushm->count*/total_num,
-        12 + 14 - strlen(datemsg) + (ufo / 60 < 10),
+        (int)(unsigned int)(12 + 14 - strlen(datemsg) + (ufo / 60 < 10)),
         cuser.userid, flagmsg);
     outf(footer);
 }
